agent/pyproject.toml
[project]
name = "omnitool-agent"
version = "0.1.0"
description = "AI agent with Gradio UI - orchestrates desktop control via parser + desktop servers"
requires-python = ">=3.12,<3.13"
dependencies = [
    "gradio>=6.9.0",
    "anthropic>=0.84.0",
    "openai",
    "groq>=1.1.1",
    "requests>=2.32.5",
    "pillow>=12.1.1",
    "pyngrok",
]
